rename the plugin and migrate to J5 Fixes #1
