+++ README.md
@@ -1,5 +1,6 @@
 
-<h2>Plugin_Field_File</h2>
+<h2>ITCS_Field_File</h2>
+<p><small>The old name ist plg_field_file</small></p>
 <p>A custom field based on the Media Manager with which files (documents, videos and audio) are provided.</p>
 <p><img src="https://j4.it-conserv.de/images/plugins/plugin_field_file/plg_field_file_preview1.jpg" alt="example" width="600"></p>
 <p>Further downloads and information are available at <a href="https://it-conserv.de">https://it-conserv.de</a></p>
