+++ languages/de-DE/plg_fields_file.ini
@@ -5,14 +5,14 @@
 ; License - http://www.gnu.org/licenses/gpl-3.0.html GNU/GPL
 ; Websites: it-conserv.de
 
-PLG_FIELDS_FILE="Fields - Datei"
-PLG_FIELDS_FILE_XML_DESCRIPTION="Mit diesem Plugin können Sie neue Felder vom Typ \"Datei\" in allen Erweiterungen erstellen, in denen benutzerdefinierte Felder unterstützt werden. Dieses Plugin verwendet den Medienmanager zum Auswählen und Hochladen einer neuen Datei."
-PLG_FIELDS_FILE_LABEL="Datei (%s)"
-PLG_FIELDS_FILE_PARAMS_DIRECTORY_LABEL="Verzeichnis"
-PLG_FIELDS_FILE_FILENAME_LABEL="Dateiname"
-PLG_FIELDS_FILE_PREVIEW_LABEL="Vorschau Modal"
-PLG_FIELDS_FILE_DOWNLOAD_LABEL="Download Button"
-PLG_FIELDS_FILE_RAW_LABEL="Dateipfad"
-PLG_FIELDS_FILE_PDF_VIEWER_MISSING="<p>Kein PDF Viewer vorhanden. <a href=\"%s\">Hier klicken zum Download des PDF Dokument.</a></p>"
+ITCS_FIELDS_FILE="Felder - Datei"
+ITCS_FIELDS_FILE_XML_DESCRIPTION="Mit diesem Plugin können Sie neue Felder vom Typ \"Datei\" in allen Erweiterungen erstellen, in denen benutzerdefinierte Felder unterstützt werden. Dieses Plugin verwendet den Medienmanager zum Auswählen und Hochladen einer neuen Datei."
+ITCS_FIELDS_FILE_LABEL="Datei (%s)"
+ITCS_FIELDS_FILE_PARAMS_DIRECTORY_LABEL="Verzeichnis"
+ITCS_FIELDS_FILE_FILENAME_LABEL="Dateiname"
+ITCS_FIELDS_FILE_PREVIEW_LABEL="Vorschau Modal"
+ITCS_FIELDS_FILE_DOWNLOAD_LABEL="Download Button"
+ITCS_FIELDS_FILE_RAW_LABEL="Dateipfad"
+ITCS_FIELDS_FILE_PDF_VIEWER_MISSING="<p>Kein PDF Viewer vorhanden. <a href=\"%s\">Hier klicken zum Download des PDF Dokument.</a></p>"
 PREVIEW="Vorschau"
 DOWNLOAD="Download"
