+++ languages/en-GB/plg_fields_file.sys.ini
@@ -5,13 +5,13 @@
 ; License - http://www.gnu.org/licenses/gpl-3.0.html GNU/GPL
 ; Websites: it-conserv.de
 
-PLG_FIELDS_FILE="Fields - File"
-PLG_FIELDS_FILE_XML_DESCRIPTION="This plugin lets you create new fields of type 'file' in any extensions where custom fields are supported. This plugin uses the media manager to select and upload a new file."
-PLG_FIELDS_FILE_LABEL="File (%s)"
-PLG_FIELDS_FILE_PARAMS_DIRECTORY_LABEL="Directory"
-PLG_FIELDS_FILE_FILENAME_LABEL="Filename"
-PLG_FIELDS_FILE_PREVIEW_LABEL="Preview Modal"
-PLG_FIELDS_FILE_DOWNLOAD_LABEL="Download Button"
-PLG_FIELDS_FILE_RAW_LABEL="Filepath"
+ITCS_FIELDS_FILE="Fields - File"
+ITCS_FIELDS_FILE_XML_DESCRIPTION="This plugin lets you create new fields of type 'file' in any extensions where custom fields are supported. This plugin uses the media manager to select and upload a new file."
+ITCS_FIELDS_FILE_LABEL="File (%s)"
+ITCS_FIELDS_FILE_PARAMS_DIRECTORY_LABEL="Directory"
+ITCS_FIELDS_FILE_FILENAME_LABEL="Filename"
+ITCS_FIELDS_FILE_PREVIEW_LABEL="Preview Modal"
+ITCS_FIELDS_FILE_DOWNLOAD_LABEL="Download Button"
+ITCS_FIELDS_FILE_RAW_LABEL="Filepath"
 PREVIEW="Preview"
 DOWNLOAD="Download"
